turned DisplayROB into a component as it should've been
