+++ components/ROBDebugger.tsx
@@ -1,203 +1,17 @@
 import React, { useState } from "react";
-import * as Utils from "@/lib/utils";
+import {
+  extractSignalValue,
+  extractSignalValueToInt,
+  parseROBData,
+} from "@/lib/utils";
 import * as Constants from "@/lib/constants";
 import * as Types from "@/lib/types";
-
-type SignalType = {
-  sigType: string;
-  width: number;
-};
-
-type SignalData = {
-  name: string;
-  type: SignalType;
-  value: string;
-};
-
-type ScopeData = {
-  [key: string]: SignalData | { children: ScopeData };
-};
+import { SignalType, SignalData, ScopeData } from "@/lib/tstypes";
+import DisplayROBData from "./DisplayROBData";
 
 type ROBDebuggerProps = {
   className: string;
   signalData: ScopeData;
-};
-
-const parseROBData = (entries: string, arrLen: number): Types.ROB_DATA[] => {
-  // Remove the 'b' prefix if present
-  const binaryStr = entries.startsWith("b") ? entries.slice(1) : entries;
-
-  const result: Types.ROB_DATA[] = [];
-  const entryWidth = Types.ROB_DATA_WIDTH;
-
-  // Helper function to extract bits and convert to number
-  const extractBits = (startIndex: number, length: number): number => {
-    const bitsSlice = binaryStr.slice(startIndex, startIndex + length);
-    return parseInt(bitsSlice, 2);
-  };
-
-  // Process each ROB entry from the end to the beginning
-  for (let i = arrLen - 1; i >= 0; i--) {
-    const startIdx = i * entryWidth;
-
-    // Extract fields from left to right in the entry
-    const T_old = extractBits(startIdx, Types.PHYS_REG_TAG_WIDTH);
-
-    const T_new = extractBits(
-      startIdx + Types.PHYS_REG_TAG_WIDTH,
-      Types.PHYS_REG_TAG_WIDTH
-    );
-
-    const R_dest = extractBits(
-      startIdx + 2 * Types.PHYS_REG_TAG_WIDTH,
-      Types.REG_IDX_WIDTH
-    );
-
-    const valid =
-      binaryStr[
-        startIdx + 2 * Types.PHYS_REG_TAG_WIDTH + Types.REG_IDX_WIDTH
-      ] === "1";
-
-    const retireable =
-      binaryStr[
-        startIdx + 2 * Types.PHYS_REG_TAG_WIDTH + Types.REG_IDX_WIDTH + 1
-      ] === "1";
-
-    result.push({
-      T_old,
-      T_new,
-      R_dest,
-      valid,
-      retireable,
-    });
-  }
-
-  return result;
-};
-
-const DisplayROBData = (
-  ROBData: Types.ROB_DATA[],
-  head: number,
-  tail: number,
-  isROB: boolean
-) => {
-  return (
-    <>
-      <table className="">
-        <thead className="border ROB-border-color bg-slate-300">
-          <tr>
-            <th className=" p-2">Entry #</th>
-            <th className=" border-l ROB-border-color p-2">R_dest</th>
-            <th className=" border-l ROB-border-color p-2">T_new</th>
-            <th className=" border-l ROB-border-color p-2">T_old</th>
-            <th className=" border-l ROB-border-color p-2">Valid</th>
-            {isROB && (
-              <>
-                <th className=" border-l ROB-border-color p-2">Retirable</th>
-                <th className=" border-l ROB-border-color p-2 pl-9">
-                  Head/Tail
-                </th>
-              </>
-            )}
-          </tr>
-        </thead>
-        <tbody>
-          {ROBData.map((entry, idx) => {
-            let isHead = head === idx;
-            let isTail = tail === idx;
-
-            if (!isROB) {
-              isHead = false;
-              isTail = false;
-            }
-
-            const isBoth = isHead && isTail;
-            const isEither = isHead || isTail;
-
-            const entryNumber = idx.toString().padStart(2, "") + ":";
-
-            // Helper function to display the value or "NaN"
-            const displayValue = (value: any) => (isNaN(value) ? "XX" : value);
-
-            // green if tail, red if head, yellow if both
-            const color = isBoth
-              ? "bg-yellow-200"
-              : isHead
-              ? "bg-red-200"
-              : isTail
-              ? "bg-green-200"
-              : entry.valid
-              ? "bg-yellow-100"
-              : "bg-gray-200";
-            const headOrTailString =
-              "←" +
-              (isBoth
-                ? " Head/Tail"
-                : isHead
-                ? " Head"
-                : isTail
-                ? " Tail"
-                : "");
-
-            return (
-              <tr key={idx} className={`border ROB-border-color ${color}`}>
-                <td className="text-right border-l ROB-border-color">
-                  {entryNumber}
-                </td>
-                <td className="text-center border-l ROB-border-color">
-                  {"r" + displayValue(entry.R_dest)}
-                </td>
-                <td className="text-center border-l ROB-border-color">
-                  {"p" + displayValue(entry.T_new)}
-                </td>
-                <td className="text-center border-l ROB-border-color">
-                  {"p" + displayValue(entry.T_old)}
-                </td>
-                <td className="text-center border-l ROB-border-color">
-                  {displayValue(entry.valid ? "1" : "0")}
-                </td>
-                {isROB && (
-                  <>
-                    <td className="text-center border-l ROB-border-color">
-                      {displayValue(entry.retireable ? "1" : "0")}
-                    </td>
-                    <td className="text-center border-l border-r ROB-border-color pl-1">
-                      {isEither && headOrTailString}
-                    </td>
-                  </>
-                )}
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
-    </>
-  );
-};
-
-// function to extract values from the signal data
-const extractSignalValue = (
-  signalData: ScopeData,
-  name: string
-): SignalData => {
-  return (signalData as unknown as { children: ScopeData }).children[
-    name
-  ] as SignalData;
-};
-
-const extractSignalValueToInt = (
-  signalData: ScopeData,
-  name: string
-): number => {
-  let signalValue = (
-    (signalData as unknown as { children: ScopeData }).children[
-      name
-    ] as SignalData
-  ).value;
-
-  return signalValue.startsWith("b")
-    ? Utils.binaryStringToInt(signalValue)
-    : parseInt(signalValue);
 };
 
 const ROBDebugger: React.FC<ROBDebuggerProps> = ({ className, signalData }) => {
@@ -236,7 +50,13 @@ const ROBDebugger: React.FC<ROBDebuggerProps> = ({ className, signalData }) => {
             ROB (size = {Constants.ROB_SZ})
           </h2>
           {/* display inputs */}
-          {DisplayROBData(ROB_dispatched_ins, -1, -1, false)}
+          <DisplayROBData
+            className=""
+            ROBData={ROB_dispatched_ins}
+            head={-1}
+            tail={-1}
+            isROB={false}
+          />
 
           {/* display ROB internals */}
           <div className="flex space-x-4 mt-4">
@@ -268,7 +88,13 @@ const ROBDebugger: React.FC<ROBDebuggerProps> = ({ className, signalData }) => {
           </div>
 
           {/* display ROB entries */}
-          {DisplayROBData(ROB_entries, head, tail, true)}
+          <DisplayROBData
+            className=""
+            ROBData={ROB_entries}
+            head={head}
+            tail={tail}
+            isROB={true}
+          />
 
           {/* output signals */}
           <div>
